fix: moved tests to admin package
